Prepared support for new build on Mac OS X Time out for not existing directories after 10 seconds Check directory existence in parallel Inform about all missing directories via a single error message Show remaining time considering relative error of 10% Check for grid icon updates only when needed Revised directory lock process detection Implemented high resolution icons Accessibility: fixed unreadable labels More polished user interfaces Fixed time stamp not being set on NFS/Samba shares (Linux)

+++ zen/file_traverser.cpp
@@ -17,7 +17,7 @@
 #include "dll.h"
 #include "FindFilePlus/find_file_plus.h"
 
-#elif defined FFS_LINUX
+#elif defined FFS_LINUX || defined FFS_MAC
 #include <sys/stat.h>
 #include <dirent.h>
 #endif
@@ -498,8 +498,7 @@ private:
     const DWORD volumeSerial; //may be 0!
 };
 
-
-#elif defined FFS_LINUX
+#elif defined FFS_LINUX || defined FFS_MAC
 class DirTraverser
 {
 public:
@@ -515,7 +514,7 @@ public:
                    the buffer whose address is passed in entry as follows:
                        len = offsetof(struct dirent, d_name) + pathconf(dirpath, _PC_NAME_MAX) + 1
                        entryp = malloc(len); */
-        const long maxPath = std::max<long>(::pathconf(directoryFormatted.c_str(), _PC_NAME_MAX), 10000); //::pathconf may return -1
+        const size_t maxPath = std::max<long>(::pathconf(directoryFormatted.c_str(), _PC_NAME_MAX), 10000); //::pathconf may return long(-1)
         buffer.resize(offsetof(struct ::dirent, d_name) + maxPath + 1);
 
         traverse(directoryFormatted, sink, 0);
